requirements.txt
flask~=3.1.1
flask-sqlalchemy
flask-migrate
flask-login
psycopg2-binary
python-dotenv~=1.1.1
werkzeug~=3.1.3
wtforms~=3.2.1
email_validator
SQLAlchemy~=2.0.41
requests~=2.32.4
urllib3~=2.5.0
alembic~=1.16.2